backend/requirements.txt
django
djangorestframework
django-cors-headers
psycopg[binary]
httpx
gunicorn
pytest
pytest-django
